requirements.txt
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.7.2
gunicorn==19.9.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
whitenoise==4.0
django-ckeditor==5.6.1
django-import-export==1.0.1
celery==4.2.1
WeasyPrint==0.42.3